#2696 - Document-level recommendations
- Move visibility calculation code into the layer recommender supports
